chore: F5 housekeeping one-liners (comment/doc drift + util rename)
Three unrelated housekeeping fixes bundled per the backlog F5 lane:

1. scripts/generate-brand-assets.mjs: the header comment listed
   Montserrat-{Bold,SemiBold,Italic}.ttf, but the code loads
   Montserrat-MediumItalic.ttf (:28). Corrected the stale Italic entry.

2. CLAUDE.md: RecipeContext.tsx no longer exists (only AuthContext.tsx
   remains under src/context/), yet two references still described it as
   'commented out'. Updated both to say it was removed. Also refreshed the
   src/util listing to name formatQuantity (see #3).

3. Renamed src/util/validateIngredientQuantityStr.ts -> formatQuantity.ts.
   The file's validation role is gone (parsing moved to
   @jclind/ingredient-parser); its only export is the closestFraction
   display formatter, so the name now matches. Updated all 5 import sites
   (IngredientItemText, SingleRecipe, PrintableRecipe, closestFraction.test,
   and the commented ref in updateIngredients).

Gates: tsc clean, Vitest 556 passed / 2 skipped (75 files), build clean.
